Fixes #4 Add new folder structure to samples.
This commit changes the structure of the samples folder.
Experimental: This is for samples that are not yet ready and don't have to be valid json payloads when validated against the schema.
Proposed: This folder holds valid samples that are yet to be approved. The payloads here should be valid.
Approved: This holds the official samples. Once the samples in the proposed folder are approved, it's just a matter of moving them to the approved folder.
The schema validation script will only run against the proposed and approved folder structures.