Add UWP card visualizer
This change adds the UWP application that will eventually load json and
use the UWP card renderer to display the generated xaml/image.  The
visualizer in this change wires up the current renderer and displays
it's xaml output, which is hardcoded to a blue square.

Also included in this change are the code changes to allow the renderer
component to be instantiated from its winmd, including proper calling of
mdmerge and the hinting in the consuming csproj.