Dev cwhytock (#78)
* Initial framework for the UWP Xaml renderer

This change add the basic framework of the UWP Xaml renderer component.
Ignore the actual methods defined in the idl, they exist to enable
hooking up and calling from a future test application.

* Remove accidentally committed file

AdaptiveCards.XamlCardRenderer.h is generated from the idl, so the
header itself shouldn't be checked in.

* Update component version to NTDDI_WIN10_RS1

Update component version to NTDDI_WIN10_RS1

* Update renderer to current dev spec

Update renderer to current dev spec

* Add UWP card visualizer

This change adds the UWP application that will eventually load json and
use the UWP card renderer to display the generated xaml/image.  The
visualizer in this change wires up the current renderer and displays
it's xaml output, which is hardcoded to a blue square.

Also included in this change are the code changes to allow the renderer
component to be instantiated from its winmd, including proper calling of
mdmerge and the hinting in the consuming csproj.