XamlTileRenderer:Render Image elements
This change adds the basic rendering of image element to the xaml card
renderer.  While the element creation follows the same pattern as the
existing text blocks, there's a chunk of extra logic dedicated to
tracking the actual loading of the image pixels.  While not currently
used, this will be used by task 10826542 to ensure all images are loaded
before rendering the xaml tree to a single image.