XamlTileRenderer:Delay rendering completion until images are fully available (#109)
This change creates a RenderCardAsXamlAsync that returns an async
operation
which only completes once all the images have been downloaded and are
ready
to be displayed.

There are a couple of main pieces to this changes:
1) Image loading is now done manually instead of letting
BitmapImage::SetSource do it.  This is
because Xaml delays image loading until the control is being rendered in
the tree.  Since that
will never happen, the ImageLoaded event neve fires it the source is set
this way.
2) I created a common async operation base class which will be reused to
provide the same functionality
for the implementation of the RenderCardAsImageAsync operation.
3) Both the ImageLoadTracker and the XamlBuilder now expose a callback
interface to notify clients
when all images have been loaded.

I also implemented a few bits of the default resource dictionary and the
usage of resource, just to
verify that strategy will work.