Merge Dev into Master test. (#170)
* Initial framework for the UWP Xaml renderer (#72)

* Initial framework for the UWP Xaml renderer

This change adds the basic framework of the UWP Xaml renderer component.
Ignore the actual methods defined in the idl, they exist to enable
hooking up and calling from a future test application.

* Remove accidentally committed file

AdaptiveCards.XamlCardRenderer.h is generated from the idl, so the
header itself shouldn't be checked in.

* Update component version to NTDDI_WIN10_RS1

Update component version to NTDDI_WIN10_RS1

* Add TextBlock to model. (#73)

* Add TextBlock to model and renderer

* Dev cwhytock (#78)

* Initial framework for the UWP Xaml renderer

This change add the basic framework of the UWP Xaml renderer component.
Ignore the actual methods defined in the idl, they exist to enable
hooking up and calling from a future test application.

* Remove accidentally committed file

AdaptiveCards.XamlCardRenderer.h is generated from the idl, so the
header itself shouldn't be checked in.

* Update component version to NTDDI_WIN10_RS1

Update component version to NTDDI_WIN10_RS1

* Update renderer to current dev spec

Update renderer to current dev spec

* Add UWP card visualizer

This change adds the UWP application that will eventually load json and
use the UWP card renderer to display the generated xaml/image.  The
visualizer in this change wires up the current renderer and displays
it's xaml output, which is hardcoded to a blue square.

Also included in this change are the code changes to allow the renderer
component to be instantiated from its winmd, including proper calling of
mdmerge and the hinting in the consuming csproj.

* Additions to object model. (#77)

* Restructure Project
Add Parser.
Split library.
Add additional properties to object model.

* UWP renderer textblock support (#90)

* XamlTileRenderer:Render Text elements

Adds renderer support for processing and creating TextBlock elements.
As this
is the first control type supported, this also adds some of the general
code paths
and helper functions that will be used by other controls.

* Code review feedbacl

CR feedback on the previous commit

* Add Image visual properties to object model (#92)

* Add Image visual properties to object model
Add Container properties
Add missing TextBlock property

Add Image projection
Change Image Uri to Uri class instead of string

* Fix Container constructor because android does not like the default parameter vector

* XamlTileRenderer:Render Image elements (#98)

This change adds the basic rendering of image element to the xaml card
renderer.  While the element creation follows the same pattern as the
existing text blocks, there's a chunk of extra logic dedicated to
tracking the actual loading of the image pixels.  While not currently
used, this will be used by task 10826542 to ensure all images are loaded
before rendering the xaml tree to a single image.

* enabling image parsing (#102)

* Add Container Projection (#104)

* Add Container Projection

* Fix formatting

* Implement size property

* fix final formatting error

* XamlTileRenderer:Delay rendering completion until images are fully available (#109)

This change creates a RenderCardAsXamlAsync that returns an async
operation
which only completes once all the images have been downloaded and are
ready
to be displayed.

There are a couple of main pieces to this changes:
1) Image loading is now done manually instead of letting
BitmapImage::SetSource do it.  This is
because Xaml delays image loading until the control is being rendered in
the tree.  Since that
will never happen, the ImageLoaded event neve fires it the source is set
this way.
2) I created a common async operation base class which will be reused to
provide the same functionality
for the implementation of the RenderCardAsImageAsync operation.
3) Both the ImageLoadTracker and the XamlBuilder now expose a callback
interface to notify clients
when all images have been loaded.

I also implemented a few bits of the default resource dictionary and the
usage of resource, just to
verify that strategy will work.

* Add columnSet and Column projections (#111)

* Add columnSet and Column projections

* CR Comments

* CR Comments V2

* add entries to manifest file

* Dev cwhytock uwpjsonparsing (#118)

* XamlTileRenderer:Integrate JSON parser

This change projects the parser through the projection object model and
provides component
methods to perform direct json -> Xaml/image.

I also flushed out some more of the styling of text elements.

* Code review feedback

code review feedback

* Cwhytock dev componentize (#146)

* XamlTileRenderer:Render Container elements

Implement support for the ColumnSet primitive for the UWP Renderer.

* Componnet refactor

Refactor the UWP renderer appropriately

* Fix solution build

* Update schema pt 1: Shared Object Model (#149)

* Update Shared Object Model to current schema specifications

* CR Comments

* Update UWP projections and Renderer to current schema (#156)

* Update UWP Projections to match current schema

* Update Renderer to current schema

* CR Comments

* Fix SAL annotations

* Move Shared and UWP files to new structure

* Get everything UWP building again.

* sln fixes