Fix UWP padding in first column
UWP was accidently adding a column spacer before the first column in
column sets, resulting in incorrect padding appearing on the left of the
first column, as described in issue #340

Simply added an if check that skips adding the spacer if a column hasn't
been added yet. Tested and verified changes using Visualizer.