Fix UWP person images that don't have size set (#355)
In the UWP renderer, if you're using an image with "person" style, and
you don't manually specify the size of the image, it was failing to
render, as described in issue #331. Non-person images render fine
without the style provided.

Code was forgetting to set the stretch behavior when image size was set
to Default inside the person logic branch of the code. Just had to add
that line so it included Default size.