UWP: Added DataJson property to Action.Submit
Added the DataJson property to the UWP object model. DataJson returns a string of the JSON, consistent with the .NET version of the library that has a DataJson property.

Tested with visualizer app and Notifications Visualizer, confirmed that the "data" property is now being passed through and accessible by renderer hosts.