Add backward compatibility for "separation" and "Column.size" (#505)
* Some minor fixes, and introduce actionStyle
It is now possible to explicitly select the button style ("link" or "button") in ActionSet elements and on the card itself.

* Fix ChoiceSetInput.value throwing in unrendered ShowCardAction cards

* Add a way to pre-expand a single ShowCardAction

* New spacing model
Removes separation and introduces spacing/separator
See issue https://github.com/Microsoft/AdaptiveCards/issues/462

* Small fix for Image (make size "auto" by default)
Some cleanup

* Add backward compat for "separation" and "Column.size"
To make things easier, the renderer will for now accept both the old "separation" and new "spacing" as well as the old "Column.size" and new "Column.width". When the old properties are used, the visualizer will display an error indicating that the property has been deprecated and will be removed.