Add methods to easily work with elements by Id
This adds a few helper methods to make it easier to work with elements in a card, especially when loaded from JSON. The GetElementById method traverses the Card's tree and returns the first element with the Id. The SetImage and SetText methods make it easy to change the Url of an Image, the Text of a TextBlock or the Value of a TextInput anywhere in the Card. These methods are especially useful for authoring Cards in JSON and modifying the content in code.