Fix AutoImageSet inconsistencies when loading images asynchronously.
We had logic to make sure that we don't stretch an image when set to autosize.
The logic would check the size of the parent and set the MaxWidth and MaxHeight if the parent was bigger.
When in a ImageSet, the parent would be the whole grid and that logic would not work properly.
This changes the logic to simply force the MaxWidth and MaxHeight of Auto images to their pixel size.

I've also added a new comparison on the TestApp to show the live XAML card.
This is the card before being rendered to a bitmap.
It can then be examined to see the root cause of test failures using the live tree viewer.