Spec tooling and specs (#3068)
Added tooling so that we can generate specs from our schema, so that our spec'ing process starts with updating the schema first, leading to less errors and mistakes in the end :)

In `schemas\src`, add your new properties/elements to the schema

![UpdateSchema](https://user-images.githubusercontent.com/13246069/59529888-0fee6b80-8e97-11e9-8925-d29264c57b4d.gif)


### 3. Generate the updated spec

If you followed the setup instructions, simply save the file and the spec will be updated! Add any necessary rendering details to the spec.

![UpdatedSpec](https://user-images.githubusercontent.com/13246069/59530207-e2ee8880-8e97-11e9-8957-de92911c2935.gif)


## Review spec

Submit a pull request for your updated spec/schema. The proposed schema in the proposal becomes outdated and should be replaced with a link to the PR.

![image](https://user-images.githubusercontent.com/13246069/59530515-a8392000-8e98-11e9-8d25-62d40d550517.png)


I'll share more details about this at Monday's meeting. Lots of details about the new proposed spec'ing process are captured here: https://github.com/microsoft/AdaptiveCards/wiki/Project-Management

###### Microsoft Reviewers: [Open in CodeFlow](http://wpcp.azurewebsites.net/CodeFlowProtocolProxy2.php?pullrequest=https://github.com/microsoft/AdaptiveCards/pull/3068)