Remove symbols nuget package signing reference (#3596)
The UWP renderer is native code which is unsupported by the nuget symbol server. 

If you nuget pack without the -Symbols flag then the pdb's will be bundled into the regular package, so consumers can still get the pdbs. However, the symbols.nupkg will never be generated.

This config is used as input after the symbols package is created so shouldn't reference a package that won't exist.